+++ containers/forms.py
@@ -28,9 +28,14 @@ class ContainerForm(forms.ModelForm):
         super().__init__(*args, **kwargs)
         # Hide project field
         self.fields["project"].widget = forms.HiddenInput()
+
         # Hide host port if in docker-shared mode
         if settings.KIOSC_NETWORK_MODE == "docker-shared":
             self.fields["host_port"].widget = forms.HiddenInput()
+
+        # Make host port field mandatory in host mode
+        if settings.KIOSC_NETWORK_MODE == "host":
+            self.fields["host_port"].required = True
 
     def clean(self):
         """Override to check for secret keys in the environment."""