+++ containers/tests/factories.py
@@ -42,8 +42,7 @@ class ContainerFactory(factory.django.DjangoModelFactory):
     container_port = 80
     container_path = ""
     heartbeat_url = ""
-    # host_port = factory.Sequence(lambda n: 8000 + n)
-    host_port = None
+    host_port = factory.Sequence(lambda n: 8000 + n)
     timeout = 60
     state = STATE_INITIAL
     environment = "{}"